requirements.txt
absl-py==1.4.0
addict==2.4.0
aniso8601==9.0.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
arrow==1.2.3
black==22.3.0
cachetools==5.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
Cython==0.29.34
-e git+https://github.com/facebookresearch/detectron2.git@96c752ce821a3340e27edd51c28a00665dd32a30#egg=detectron2
fairscale==0.4.6
filelock==3.12.0
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.38.0
fsspec==2023.1.0
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.20.0
google-auth-oauthlib==0.4.6
grpcio==1.55.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.1
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lap==0.4.0
Levenshtein==0.21.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdurl==0.1.2
mmcv==1.6.0
mmdet==3.0.0
mmengine==0.7.3
model-index==0.1.11
motmetrics==1.4.0
mypy-extensions==1.0.0
numpy==1.21.6
nvgpu==0.10.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==1.3.5
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.3
polygon==1.1.0
Polygon3==3.0.9.1
portalocker==2.7.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pycocotools==2.0.6
pydot==1.4.2
Pygments==2.15.1
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.1.1
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
safetensors==0.3.1
scipy==1.7.3
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
timm==0.9.2
tomli==2.0.1
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
tqdm==4.65.0
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==2.0.2
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
xmltodict==0.13.0
yacs==0.1.8
yapf==0.33.0
zipp==3.15.0
